Re-lock dependencies on a Raspberry Pi
I couldn't install this on a Pi, because it incorrectly chose
versions of e.g. numpy that did not exist in binary form for the
pi. I re-locked dependencies with:

`PIP_ONLY_BINARY=":all:" pipenv lock --dev`

This forced the use of wheels, and not only completed but did so
in ~10 minutes.