Generate zip files before download
Previously, zip files were only downloaded if they already
existed. Now, they are created with  an action, then
downloaded from the result.

Chrome now insists on downloading the files as `output.zip`, I can't see why this has changed, but perhaps it's related
to the security warning. This seems to be specific to zip files.