Add a timeout when shutting down the server.
The browser needs to be closed for graceful shutdown. If the browser
is left open the microscope may never shut down as it is waiting on
the uvicorn process to finish. This causes potential SD card issues.

This forces the server to force shutdown if it doesn't exit gracefully in
2 seconds.