Split up the action button methods for monitoring and starting a task
Previously the start task method of action button handled awaiting
polling and resetting variables when the task completes. If an ongoing
task was found polling was started so that logs could be communicated
but no clean up happened when the task finished. This has been split
up and the checks for ongoing tasks improved so they also clean up
after ongoing tasks finish.

Closes #442